Split git-delta into a binary and library
